Fix wrong type in encoder_update

+++ keyboards/planck/keymaps/lereverandnox/user_functions.c
@@ -135,7 +135,7 @@ uint16_t get_tapping_term(uint16_t keycode) {
 }
 
 void encoder_update_user(uint8_t index, bool clockwise) {
-    static uint8_t kc;
+    static uint16_t kc;
     uint8_t mods = get_mods();
     if (clockwise) {
         if ((mods & MOD_MASK_CTRL) && (mods & MOD_MASK_GUI) && (mods & MOD_MASK_ALT) && (mods & MOD_MASK_SHIFT)) { // CTRL + GUI + ALT + SHIFT
